비밀방 설정 변경

+++ src/main/java/egovframework/example/sample/web/GoldRoom.java
@@ -2,7 +2,7 @@ package egovframework.example.sample.web;
 
 public class GoldRoom {
     public int maxplayer;
-    public int ante;
+    public long ante;
     public long maxbetvalue;
     public boolean isprivate;
 }
